refactor(modal): 모달 타입 수정 및 child props 생성
